+++ src/index.js
@@ -1,6 +1,6 @@
 import './assets/global/reset.css';
 import './assets/main.scss';
 
-import App from './app';
+import controller from './controller';
 
-App();
+controller();
